requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
google-generativeai
pandas
pydantic
requests
sqlalchemy
psycopg2-binary
pymysql
langchain-google-genai
langchain-community
langchain-core
sqlparse
